fix(web): refine authorization errors and pairing accessibility
