+++ apps/web/src/cloud/connectCliAuth.test.ts
@@ -108,19 +108,14 @@ describe("connectCliAuth", () => {
     ).toBeNull();
   });
 
-  it("reads the OAuth error Clerk sends when sign-in is refused", () => {
+  it("reads the OAuth error code Clerk sends when sign-in is refused", () => {
     expect(
       readConnectCliCallbackError(
         new URL(
           "https://app.t3.codes/connect/callback?error=access_denied&error_description=The+user+denied+the+request",
         ),
       ),
-    ).toEqual({ error: "access_denied", description: "The user denied the request" });
-    expect(
-      readConnectCliCallbackError(
-        new URL("https://app.t3.codes/connect/callback?error=server_error&state=s"),
-      ),
-    ).toEqual({ error: "server_error", description: null });
+    ).toBe("access_denied");
     expect(
       readConnectCliCallbackError(
         new URL("https://app.t3.codes/connect/callback?code=abc&state=state-1"),
