+++ apps/web/src/components/T3Wordmark.tsx
@@ -1,4 +1,7 @@
-import type { SVGProps } from "react";
+import type { ComponentProps, SVGProps } from "react";
+
+import { cn } from "~/lib/utils";
+import { Badge } from "./ui/badge";
 
 export function T3Wordmark(props: SVGProps<SVGSVGElement>) {
   return (
@@ -8,5 +11,43 @@ export function T3Wordmark(props: SVGProps<SVGSVGElement>) {
         fill="currentColor"
       />
     </svg>
+  );
+}
+
+/**
+ * The wordmark with its "Code" label, as the sidebar header and the standalone
+ * auth pages show it. `onBackdrop` switches to the white treatment used over
+ * release-channel artwork.
+ */
+export function T3CodeWordmark({ onBackdrop = false }: { readonly onBackdrop?: boolean }) {
+  return (
+    <span
+      className={cn(
+        "inline-flex min-w-0 items-baseline gap-1",
+        onBackdrop ? "text-white" : "text-foreground",
+      )}
+    >
+      <T3Wordmark aria-label="T3" className="h-2.5 w-auto shrink-0" />
+      <span
+        className={cn(
+          "truncate text-sm font-medium tracking-tight",
+          onBackdrop ? "text-white/70" : "text-muted-foreground",
+        )}
+      >
+        Code
+      </span>
+    </span>
+  );
+}
+
+/** Release-channel pill ("Dev", "Nightly") shown beside the wordmark. */
+export function EnvironmentStagePill({ className, ...props }: ComponentProps<typeof Badge>) {
+  return (
+    <Badge
+      className={cn("ml-1 rounded-full px-1.5 text-muted-foreground", className)}
+      size="sm"
+      variant="secondary"
+      {...props}
+    />
   );
 }
