+++ apps/web/src/components/sidebar/SidebarChrome.tsx
@@ -11,7 +11,7 @@ import { Link, useCanGoBack, useLocation, useNavigate } from "@tanstack/react-ro
 import { useEnvironmentIdentificationMode } from "../../hooks/useSettings";
 import { cn } from "../../lib/utils";
 import { useEnvironments } from "../../state/environments";
-import { EnvironmentStagePill, T3CodeBrand } from "../T3CodeBrand";
+import { EnvironmentStagePill, T3CodeWordmark } from "../T3Wordmark";
 import {
   resolveEnvironmentIdentificationPillLabel,
   resolveSidebarStageBackdropVariant,
@@ -85,7 +85,7 @@ function SidebarBrand({ onBackdrop }: { onBackdrop: boolean }) {
       className="relative z-10 ml-[var(--workspace-titlebar-content-left)] hidden h-7 w-fit min-w-0 shrink-0 items-center overflow-hidden rounded-md outline-hidden ring-ring focus-visible:ring-2 md:flex"
       to="/"
     >
-      <T3CodeBrand onBackdrop={onBackdrop} />
+      <T3CodeWordmark onBackdrop={onBackdrop} />
     </Link>
   );
 }
